1 refactored MainWindow methods across several .cpp files
The MainWindow class will be a large one, so breaking up the implementation into several files makes it easier to navigate.
