+++ src/gui/mainwindow.cpp
@@ -25,34 +25,10 @@ MainWindow::MainWindow(QSettings& settings, QWidget *parent) :
     , ui(new Ui::MainWindow)
     , m_settings(settings)
 {
-    ui->setupUi(this);
-
-    // Restore window geometry from persistent settings
-    m_settings.beginGroup("MainWindow");
-    restoreGeometry(m_settings.value("geometry").toByteArray());
-    restoreState(m_settings.value("state").toByteArray());
-    m_settings.endGroup();
-
-    // Add a label for the API version
-    ui->centralWidget->setLayout(new QHBoxLayout(ui->centralWidget));
-    ui->centralWidget->layout()->addWidget(
-        new QLabel(
-            QString::fromStdString(
-                "API version: {}"_format(api::version()))));
+    setupUi();
 }   // end constructor
 
 MainWindow::~MainWindow()
 {
     delete ui;
 }   // end destructor
-
-void MainWindow::closeEvent(QCloseEvent *event)
-{
-    // Record window geometry in persistent settings
-    m_settings.beginGroup("MainWindow");
-    m_settings.setValue("geometry", saveGeometry());
-    m_settings.setValue("windowState", saveState());
-    m_settings.endGroup();    
-
-    QMainWindow::closeEvent(event);
-}   // end closeEvent
