+++ src/gui/mainwindow/mw_events.cpp
@@ -1,0 +1,18 @@
+/**
+ * \file mw_events.cpp
+ * Implement event handlers for the `MainWindow` class
+ * 
+ * \author Igor Siemienowicz
+ * 
+ * \copyright Copyright Igor Siemienowicz 2018 Distributed under the Boost
+ * Software License, Version 1.0. (See accompanying file ../LICENSE_1_0.txt
+ * or copy at https://www.boost.org/LICENSE_1_0.txt
+ */
+
+#include "../mainwindow.h"
+
+void MainWindow::closeEvent(QCloseEvent *event)
+{
+    saveWindowGeometry();
+    QMainWindow::closeEvent(event);
+}   // end closeEvent
