+++ src/gui/mainwindow/mw_utils.cpp
@@ -1,0 +1,28 @@
+/**
+ * \file mw_utils.cpp
+ * Implement some utility methods in the `MainWindow` class
+ * 
+ * \author Igor Siemienowicz
+ * 
+ * \copyright Copyright Igor Siemienowicz 2018 Distributed under the Boost
+ * Software License, Version 1.0. (See accompanying file ../LICENSE_1_0.txt
+ * or copy at https://www.boost.org/LICENSE_1_0.txt
+ */
+
+#include "../mainwindow.h"
+
+void MainWindow::saveWindowGeometry(void)
+{
+    m_settings.beginGroup("MainWindow");
+    m_settings.setValue("geometry", saveGeometry());
+    m_settings.setValue("windowState", saveState());
+    m_settings.endGroup();    
+}   // end saveWindowGeometry method
+
+void MainWindow::restoreWindowGeometry(void)
+{
+    m_settings.beginGroup("MainWindow");
+    restoreGeometry(m_settings.value("geometry").toByteArray());
+    restoreState(m_settings.value("state").toByteArray());
+    m_settings.endGroup();
+}   // end restoreWindowGeometry method
